Add a label to a plot

This function adds a plot label to the upper left corner of a graph (or an arbitrarily specified position). It takes all the same parameters as draw_text, but has defaults that make it convenient to label graphs with letters A, B, C, etc. Just like draw_text(), it can handle vectors of labels with associated coordinates.

draw_plot_label(
  label,
  x = 0,
  y = 1,
  hjust = -0.5,
  vjust = 1.5,
  size = 16,
  fontface = "bold",
  family = NULL,
  color = NULL,
  colour,
  ...
)

Arguments

label

String (or vector of strings) to be drawn as the label.

x

The x position (or vector thereof) of the label(s).

y

The y position (or vector thereof) of the label(s).

hjust

Horizontal adjustment.

vjust

Vertical adjustment.

size

Font size of the label to be drawn.

fontface

Font face of the label to be drawn.

family

(optional) Font family of the plot labels. If not provided, is taken from the current theme.

color, colour

(optional) Color of the plot labels. If not provided, is taken from the current theme.

...

Other arguments to be handed to draw_text.