panel_border {cowplot}	R Documentation

Add/remove the panel border in a ggplot2 plot

Description

This function provides a simple way to modify the panel border in ggplot2. It doesn't do anything that can't be done just the same with theme(). However, it saves some typing.

Usage

panel_border(color = "grey85", size = 1, linetype = 1, remove = FALSE, colour)

Arguments

color, colour
The color of the border.

size
Size. Needs to be twice as large as desired outcome when panel clipping is on (the default).

linetype
Line type.

remove
If TRUE, removes the current panel border.

[Package cowplot version 1.1.1 Index]