stamp {cowplot}	R Documentation

Stamp plots with a label, such as good, bad, or ugly.

Description

Stamp plots with a label, such as good, bad, or ugly.

Usage

stamp(
  p,
  label,
  color = "black",
  alpha = 1,
  vjust = 1.1,
  hjust = 1,
  size = 14,
  family = "",
  fontface = "bold",
  clip = "on",
  colour
)

stamp_good(p, ...)

stamp_bad(p, ...)

stamp_wrong(p, ...)

stamp_ugly(p, ...)

Arguments

p
The plot to stamp

label
The text label used for the stamp

color, colour
The color of the stamp

alpha
Transparency level of the stamp

hjust, vjust
Horizontal and vertical adjustment of the label

size
Font size

family
Font family

fontface
Font face

clip
Should figure be clipped (default is "on")

...
Arguments handed off to stamp().

Examples

library(ggplot2)

p <- ggplot(iris, aes(x = Sepal.Length, y = Sepal.Width)) +
  geom_point(aes(color = factor(Petal.Width)))

stamp_bad(p + guides(color = "none"))
stamp_ugly(p)

[Package cowplot version 1.1.1 Index]