theme_cowplot {cowplot}	R Documentation

Create the default cowplot theme

Description

The default cowplot theme, with a simple half-open frame and no grid. This theme used to be set by default after loading the cowplot package, but this is no longer the case.

Usage

theme_cowplot(
  font_size = 14,
  font_family = "",
  line_size = 0.5,
  rel_small = 12/14,
  rel_tiny = 11/14,
  rel_large = 16/14
)

theme_half_open(
  font_size = 14,
  font_family = "",
  line_size = 0.5,
  rel_small = 12/14,
  rel_tiny = 11/14,
  rel_large = 16/14
)

Arguments

font_size
Overall font size.

font_family
Font family for plot title, axis titles and labels, legend texts, etc.

line_size
Line size for axis lines.

rel_small
Relative size of small text (e.g., axis tick labels)

rel_tiny
Relative size of tiny text (e.g., caption)

rel_large
Relative size of large text (e.g., title)

Details

Both theme_cowplot() and theme_half_open() provide exactly the same styling.

Value

The theme.

Examples

library(ggplot2)

ggplot(mtcars, aes(disp, mpg)) +
  geom_point() +
  theme_cowplot(font_size = 12)

[Package cowplot version 1.1.1 Index]