as_gtable {cowplot}	R Documentation

Convert plot or other graphics object into a gtable

Description

This function does its best attempt to take whatever you provide it and turn it into a gtable. It is primarily meant to convert ggplot plots into gtables, but it will also take any grid object (grob), a recorded R base plot, or a function that generates an R base plot.

Usage

as_gtable(plot)

plot_to_gtable(plot)

Arguments

plot
The plot or other graphics object to convert into a gtable. Here, plot can be any object handled by as_grob().

Details

To convert ggplot plots, the function needs to use a null graphics device. This can be set with set_null_device().

[Package cowplot version 1.1.1 Index]