draw_line {cowplot}	R Documentation

Draw a line from connected points

Description

Provide a sequence of x values and accompanying y values to draw a line on a plot.

Usage

draw_line(x, y, ...)

Arguments

x
Vector of x coordinates.

y
Vector of y coordinates.

...
geom_path parameters such as colour, alpha, size, etc.

Details

This is a convenience function, providing a wrapper around ggplot2's geom_path.

See Also

geom_path, ggdraw

Examples

ggdraw() +
  draw_line(
    x = c(0.2, 0.7, 0.7, 0.3),
    y = c(0.1, 0.3, 0.9, 0.8),
    color = "blue", size = 2
  )

[Package cowplot version 1.1.1 Index]