draw_plot {cowplot}	R Documentation

Draw a (sub)plot.

Description

Places a plot somewhere onto the drawing canvas. By default, coordinates run from 0 to 1, and the point (0, 0) is in the lower left corner of the canvas.

Usage

draw_plot(
  plot,
  x = 0,
  y = 0,
  width = 1,
  height = 1,
  scale = 1,
  hjust = 0,
  vjust = 0,
  halign = 0.5,
  valign = 0.5
)

Arguments

plot
The plot to place. Can be a ggplot2 plot, an arbitrary grob or gtable, or a recorded base-R plot, as in as_grob().

x
The x location of the plot. (Left side if hjust = 0.)

y
The y location of the plot. (Bottom side if vjust = 0.)

width
Width of the plot.

height
Height of the plot.

scale
Scales the grob relative to the rectangle defined by x, y, width, height. A setting of scale = 1 indicates no scaling.

hjust, vjust
Horizontal and vertical justification relative to x.

halign, valign
Horizontal and vertical justification of the plot inside the box.

Examples

library(ggplot2)

# make a plot
p <- ggplot(data.frame(x = 1:3, y = 1:3), aes(x, y)) +
    geom_point()
# draw into the top-right corner of a larger plot area
ggdraw() + draw_plot(p, .6, .6, .4, .4)

[Package cowplot version 1.1.1 Index]