get_panel {cowplot}	R Documentation

Retrieve the panel or part of a panel of a plot

Description

get_panel() extracts just the main panel from a ggplot or a specified panel in a faceted plot. get_panel_component() extracts components from the panel, such as geoms.

Usage

get_panel(plot, panel = NULL, return_all = FALSE)

get_panel_component(panel, pattern)

Arguments

plot
A ggplot or gtable from which to retrieve the panel

panel
An integer indicating which panel to pull. ggplot orders panels column-wise, so this is in order from the top left down.

return_all
If there is more than one panel, should all be returned as a list? Default is FALSE.

pattern
the name of the component

Value

A gtable object holding the panel(s) or a grob of the component

Examples

library(ggplot2)

p <- ggplot(mpg, aes(displ, cty)) + geom_point()
plot_panel <- get_panel(p)
ggdraw(plot_panel)

ggdraw(get_panel_component(plot_panel, "geom_point"))

[Package cowplot version 1.1.1 Index]